pyproject.toml
[project]
name = "invest-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.1",
    "httpx>=0.28.1",
    "mcp>=1.20.0",
    "openai>=2.5.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.38.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
]
